requirements.txt
discord.py
gtts
ffmpeg
asyncio
json
random
os
time
